gestor-entradas/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
apturl==0.5.2
arabic-reshaper==2.1.3
asgiref==3.5.2
asn1crypto==1.5.1
astroid==2.11.7
asttokens==2.0.8
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
Babel==2.8.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.11.1
blinker==1.4
Brlapi==0.8.3
certifi==2022.12.7
chardet==4.0.0
charset-normalizer==3.0.1
click==8.0.3
colorama==0.4.4
command-not-found==0.3
contourpy==1.0.7
cryptography==3.4.8
cssselect2==0.6.0
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.18
debugpy==1.6.3
decorator==5.1.1
defer==1.0.6
dill==0.3.5.1
distlib==0.3.4
distro==1.7.0
distro-info===1.1build1
dj-rest-auth==2.2.5
Django==4.1.7
django-active-link==0.1.8
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-crontab==0.7.1
django-environ==0.9.0
django-rest-authtoken==2.1.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docutils==0.19
duplicity==0.8.21
entrypoints==0.4
exceptiongroup==1.1.1
executing==1.1.0
fasteners==0.14.1
filelock==3.6.0
Flask==2.0.1
fonttools==4.39.3
frozenlist==1.3.3
future==0.18.2
h11==0.14.0
html5lib==1.1
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
instaloader==4.9.5
ipykernel==6.16.0
ipython==8.5.0
isort==5.10.1
itsdangerous==2.1.0
jedi==0.18.1
jeepney==0.7.1
Jinja2==3.0.3
jupyter-core==4.11.1
jupyter_client==7.3.5
keyring==23.5.0
kiwisolver==1.4.4
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy-object-proxy==1.7.1
lockfile==0.12.2
louis==3.20.0
lxml==4.9.1
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
mccabe==0.7.0
monotonic==1.6
more-itertools==8.10.0
MouseInfo==0.1.3
multidict==6.0.4
mysqlclient==2.1.1
nest-asyncio==1.5.6
netifaces==0.11.0
numpy==1.23.1
oauthlib==3.2.0
olefile==0.46
openai==0.26.5
oscrypto==1.3.0
outcome==1.2.0
packaging==21.3
pandas==1.4.3
paramiko==2.9.3
parso==0.8.3
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.5.2
prompt-toolkit==3.0.31
protobuf==3.12.4
psutil==5.9.2
ptyprocess==0.7.0
PuLP==2.6.0
pure-eval==0.2.2
PyAudio==0.2.11
PyAutoGUI==0.9.53
pycairo==1.20.1
pycodestyle==2.8.0
pycups==2.0.1
PyGetWindow==0.0.9
Pygments==2.13.0
PyGObject==3.42.1
pyHanko==0.13.2
pyhanko-certvalidator==0.19.5
pyinotify==0.9.6
PyJWT==2.6.0
pylint==2.14.5
pymacaroons==0.13.0
PyMsgBox==1.0.9
PyMySQL==1.0.2
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
PyPDF2==3.0.1
PyPDF3==1.0.6
pyperclip==1.8.2
PyRect==0.2.0
pyRFC3339==1.1
PyScreeze==0.1.28
PySocks==1.7.1
python-apt==2.4.0+ubuntu1
python-bidi==0.4.2
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
python-decouple==3.8
python3-xlib==0.15
pyttsx3==2.90
pytweening==1.0.4
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pyxdg==0.27
PyYAML==5.4.1
pyzmq==24.0.1
qrcode==7.3.1
reportlab==3.6.8
requests==2.28.2
rest-framework-simplejwt==0.0.2
SecretStorage==3.3.1
selenium==4.8.3
sentry-sdk==1.9.8
simplejson==3.17.6
six==1.16.0
sklearn==0.0.post1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SpeechRecognition==3.8.1
sqlparse==0.4.3
stack-data==0.5.1
svglib==1.3.0
systemd-python==234
tbomb==2.1.2
tinycss2==1.1.1
tomli==2.0.1
tomlkit==0.11.1
tornado==6.2
tqdm==4.64.0
traitlets==5.4.0
trio==0.22.0
trio-websocket==0.10.2
tzdata==2022.1
tzlocal==4.2
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uritools==4.0.0
urllib3==1.26.12
usb-creator==0.3.7
virtualenv==20.13.0+ds
wadllib==1.3.6
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
wrapt==1.14.1
wsproto==1.2.0
xdg==5
xhtml2pdf==0.2.8
xkit==0.0.0
xlwt==1.3.0
yarl==1.8.2
zipp==1.0.0
